cleanup parser definitions
